refactor(GoMod): Move the `Graph` class into a separate file
Improve readability and overview.
